changes in iterator

+++ src/uint256_iterator.cu
@@ -22,7 +22,8 @@ __device__ void uint256_iter::next()
     uint256_t t = curr_perm | ( curr_perm + UINT256_NEGATIVE_ONE );
 
 
-    uint8_t shift = curr_perm.ctz() + 1;
+    //uint8_t shift = curr_perm.ctz() + 1;
+    uint32_t shift = curr_perm.ctz() + 1;
 
     // add_tmp.set_all( 0x00 );
     uint256_t tmp;
